refactor(controller): 메서드 분리를 통한 책임 분리
- validation 기능은 Validator 클래스에 위임
- 하나의 책임을 갖는 메서드로 분리
