+++ src/main/java/lotto/controller/LottoController.java
@@ -1,13 +1,7 @@
 package lotto.controller;
 
-import static lotto.constants.ExceptionMessages.BONUS_NUMBER_OUT_OF_RANGE;
-import static lotto.constants.ExceptionMessages.BUY_AMOUNT_HAS_CHANGE;
-import static lotto.constants.ExceptionMessages.LOTTO_COUNT_OUT_OF_RANGE;
-import static lotto.constants.NumberConstants.LOTTO_END_NUM;
-import static lotto.constants.NumberConstants.LOTTO_START_NUM;
-import static lotto.constants.NumberConstants.MAX_TICKET_BUY_AMOUNT;
-import static lotto.constants.NumberConstants.TICKET_PRICE;
-import static lotto.constants.NumberConstants.ZERO;
+import static lotto.validator.Validator.validateBonusNumberRange;
+import static lotto.validator.Validator.validateBuyAmount;
 
 import java.util.EnumMap;
 import java.util.Map;
@@ -41,52 +35,20 @@ public class LottoController {
 
     public void run() {
 
-        // 구입 금액 입력받기
-        long buyAmount = inputHandler.readBuyAmount();
-        // TODO: validation은 다른 클래스에 책임 할당하기
-        validateLottoCount(buyAmount);
-        validateBuyAmountHasChange(buyAmount);
+        long buyAmount = getValidBuyAmount();
 
-        // 구매한 로또 번호 출력하기
         UserLotto userLotto = lottoService.getUserLotto(buyAmount);
         outputHandler.userLottoHandle(userLotto);
 
-        // 당첨 번호, 보너스 번호 입력받기
         LottoRequest lottoRequest = inputHandler.handleWinningNumber();
-        // 당첨 번호 로또 객체로 변환
         WinningLotto winningLotto = lottoService.convertWinningNumbers(lottoRequest.getWinningNumbers());
 
-        // 보너스 번호 검증
-        // TODO: validation은 다른 클래스에 책임 할당하기
-        int bonusNumber = lottoRequest.getBonusNumber();
-        validateBonusNumberRange(bonusNumber);
-        winningLotto.checkDuplicateBonusNumber(bonusNumber);
+        int bonusNumber = getValidBonusNumber(lottoRequest, winningLotto);
 
-        // 당첨 결과 계산 로직...
-        WinningResult result = new Winning().calculate(userLotto, buyAmount, winningLotto, bonusNumber);
-
-        // 당첨 결과 출력
+        WinningResult result = calculateWinningResult(userLotto, buyAmount, winningLotto, bonusNumber);
         outputHandler.winningResultHandle(result);
     }
 
-
-    private static void validateLottoCount(long buyAmount) {
-        if (buyAmount < TICKET_PRICE || buyAmount > MAX_TICKET_BUY_AMOUNT) {
-            throw new IllegalArgumentException(LOTTO_COUNT_OUT_OF_RANGE);
-        }
-    }
-
-    private static void validateBuyAmountHasChange(long buyAmount) {
-        if (buyAmount % TICKET_PRICE != ZERO) {
-            throw new IllegalArgumentException(BUY_AMOUNT_HAS_CHANGE);
-        }
-    }
-
-    private static void validateBonusNumberRange(int bonusNumber) {
-        if (bonusNumber < LOTTO_START_NUM || bonusNumber > LOTTO_END_NUM) {
-            throw new IllegalArgumentException(BONUS_NUMBER_OUT_OF_RANGE);
-        }
-    }
 
     public static Map<Rank, Integer> initializeResultMap() {
         Map<Rank, Integer> initialMap = new EnumMap<>(Rank.class);
@@ -94,5 +56,25 @@ public class LottoController {
             initialMap.put(rank, 0);
         }
         return initialMap;
+    }
+
+    private long getValidBuyAmount() {
+        long buyAmount = inputHandler.readBuyAmount();
+        validateBuyAmount(buyAmount);
+        return buyAmount;
+    }
+
+    private int getValidBonusNumber(LottoRequest lottoRequest, WinningLotto winningLotto) {
+        int bonusNumber = lottoRequest.getBonusNumber();
+        validateBonusNumberRange(bonusNumber);
+        winningLotto.checkDuplicateBonusNumber(bonusNumber);
+        return bonusNumber;
+    }
+
+    private WinningResult calculateWinningResult(UserLotto userLotto,
+                                                 long buyAmount,
+                                                 WinningLotto winningLotto,
+                                                 int bonusNumber) {
+        return new Winning().calculate(userLotto, buyAmount, winningLotto, bonusNumber);
     }
 }
